Break tests into seperate files.
